+++ crates/zstdx/src/decoding/mod.rs
@@ -13,6 +13,7 @@ pub(crate) mod decode_buffer;
 pub(crate) mod dictionary;
 pub(crate) mod flat_buffer;
 pub(crate) mod frame;
+pub(crate) mod frame_source;
 pub(crate) mod literals_section_decoder;
 #[cfg(feature = "std")]
 pub(crate) mod mt;